# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Projects RBAC API
  description: |
    REST API for authentication, groups, membership, RBAC policies, and protected assets.

    Import this file into Postman (**Import → File**), Insomnia, Thunder Client, or any OpenAPI 3 tool.
  version: 1.0.0
  contact:
    name: Projects API
servers:
  - url: http://localhost:3000
    description: Local unified Next.js app (UI + API)
  - url: "{baseUrl}"
    description: Custom / cloud deployment
    variables:
      baseUrl:
        default: https://app.example.com
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AssetResource:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        ownerId:
          type: string
        kind:
          type: string
          enum: [DOCUMENT, DASHBOARD, DATASET, SURVEY, SURVEY_RESPONSE]
        assetType:
          type: string
          enum: [Asset, Survey, SurveyResponse]
        createdBy:
          type: string
        updatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAssetRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Billing Doc
        description:
          type: string
        kind:
          type: string
          enum: [DOCUMENT, DASHBOARD, DATASET]
          default: DOCUMENT
...
    ActionLog:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
          nullable: true
        username:
          type: string
        action:
          type: string
          example: auth.login
        resourceType:
          type: string
          example: AUTH
        resourceId:
          type: string
          nullable: true
        method:
          type: string
          example: POST
        path:
          type: string
          example: /api/auth/login
        statusCode:
          type: integer
          example: 200
        success:
          type: boolean
        message:
          type: string
        ipAddress:
          type: string
        userAgent:
          type: string
        clientApp:
          type: string
        meta:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    ActionLogListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ActionLog'
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
            hasPrev:
              type: boolean
            hasNext:
              type: boolean
        sort:
          type: object
          properties:
            field:
              type: string
            order:
              type: string
              enum: [asc, desc]
paths:
...
  /api/assets:
    get:
      tags: [Assets]
      summary: List assets
      operationId: listAssets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Asset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetResource'
        '401':
          description: Unauthorized
    post:
      tags: [Assets]
      summary: Create asset
      operationId: createAsset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResource'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /api/assets/{id}:
    get:
      tags: [Assets]
      summary: Get asset by ID
      operationId: getAsset
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Asset found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResource'
        '401':
          description: Unauthorized
        '404':
          description: Not found
    put:
      tags: [Assets]
      summary: Update asset
      operationId: updateAsset
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '200':
          description: Asset updated
        '401':
          description: Unauthorized
        '404':
          description: Not found
    delete:
      tags: [Assets]
      summary: Delete asset
      operationId: deleteAsset
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Asset deleted
        '401':
          description: Unauthorized
        '404':
          description: Not found
...
  /api/logs:
    get:
      tags: [Logs]
      summary: Search and filter action logs
      description: |
        Admin-only (`LOG:READ` via admin group membership). Filtering, search, sort, and
        pagination are applied in the backend.
      operationId: listActionLogs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Free-text search across username, action, path, message, resourceId
        - in: query
          name: username
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: action
          schema:
            type: string
            example: auth.login
        - in: query
          name: resourceType
          schema:
            type: string
            example: USER
        - in: query
          name: resourceId
          schema:
            type: string
        - in: query
          name: method
          schema:
            type: string
            enum: [POST, PUT, PATCH, DELETE]
        - in: query
          name: statusCode
          schema:
            type: integer
        - in: query
          name: success
          schema:
            type: boolean
        - in: query
          name: clientApp
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: sort
          schema:
            type: string
            enum: [createdAt, username, action, resourceType, method, statusCode, success]
            default: createdAt
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        '200':
          description: Paginated action logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionLogListResponse'
        '401':
          description: Not authenticated
        '403':
          description: Forbidden
  /api/logs/filters:
    get:
      tags: [Logs]
      summary: Distinct filter values for the logs UI
      operationId: getActionLogFilters
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Filter option lists
          content:
            application/json:
              schema:
                type: object
                properties:
                  actions:
                    type: array
                    items:
                      type: string
                  resourceTypes:
                    type: array
                    items:
                      type: string
                  sortableFields:
                    type: array
                    items:
                      type: string
        '401':
          description: Not authenticated
        '403':
          description: Forbidden